+++ _output/extension_hint.php
@@ -7,6 +7,7 @@ namespace TickTackk\DeveloperTools\XF\Admin\Controller
 {
 	class XFCP_AddOn extends \XF\Admin\Controller\AddOn {}
 	class XFCP_Option extends \XF\Admin\Controller\Option {}
+	class XFCP_Phrase extends \XF\Admin\Controller\Phrase {}
 	class XFCP_Template extends \XF\Admin\Controller\Template {}
 	class XFCP_TemplateModification extends \XF\Admin\Controller\TemplateModification {}
 }